+++ edx_event_bus_kafka/internal/producer.py
@@ -18,7 +18,7 @@ from django.test.signals import setting_changed
 from openedx_events.event_bus.avro.serializer import AvroSignalSerializer
 from openedx_events.tooling import OpenEdxPublicSignal
 
-from .config import get_schema_registry_client, load_common_settings
+from .config import get_full_topic, get_schema_registry_client, load_common_settings
 
 logger = logging.getLogger(__name__)
 
@@ -157,7 +157,7 @@ def get_serializers(signal: OpenEdxPublicSignal, event_key_field: str):
     return key_serializer, value_serializer
 
 
-class EventProducerKafka():
+class KafkaEventProducer():
     """
     API singleton for event production to Kafka.
 
@@ -198,8 +198,10 @@ class EventProducerKafka():
         key_bytes = key_serializer(event_key, SerializationContext(topic, MessageField.KEY, headers))
         value_bytes = value_serializer(event_data, SerializationContext(topic, MessageField.VALUE, headers))
 
+        full_topic = get_full_topic(topic)
+
         self.producer.produce(
-            topic, key=key_bytes, value=value_bytes, headers=headers, on_delivery=on_event_deliver,
+            full_topic, key=key_bytes, value=value_bytes, headers=headers, on_delivery=on_event_deliver,
         )
 
         # Opportunistically ensure any pending callbacks from recent event-sends are triggered.
@@ -219,18 +221,16 @@ class EventProducerKafka():
         self.producer.flush(-1)
 
 
-def poll_indefinitely(api_weakref: EventProducerKafka):
+def poll_indefinitely(api_weakref: KafkaEventProducer):
     """
     Poll the producer indefinitely to ensure delivery/stats/etc. callbacks are triggered.
 
     The thread stops automatically once the producer is garbage-collected.
 
-    This ensures that callbacks are triggered in a timely fashion, rather than waiting
-    for the poll() call that we make before or after each produce() call. This may be
-    important if events are produced infrequently, and it allows the last event the
-    server emits before shutdown to have its callback run (if it happens soon enough.)
+    See ADR for more information:
+    https://github.com/openedx/event-bus-kafka/blob/main/docs/decisions/0007-producer-polling.rst
     """
-    # The reason we hold a weakref to the whole EventProducerKafka and
+    # The reason we hold a weakref to the whole KafkaEventProducer and
     # not directly to the Producer itself is that you just can't make
     # a weakref to the latter (perhaps because it's a C object.)
 
@@ -267,7 +267,7 @@ def poll_indefinitely(api_weakref: EventProducerKafka):
 # outbound-message queue and threads. The use of this cache allows the
 # producer to be long-lived.
 @lru_cache  # will just be one cache entry, in practice
-def get_producer() -> Optional[EventProducerKafka]:
+def get_producer() -> Optional[KafkaEventProducer]:
     """
     Create or retrieve Producer API singleton.
 
@@ -281,7 +281,7 @@ def get_producer() -> Optional[EventProducerKafka]:
     if producer_settings is None:
         return None
 
-    return EventProducerKafka(Producer(producer_settings))
+    return KafkaEventProducer(Producer(producer_settings))
 
 
 def on_event_deliver(err, evt):